+++ pubky-homeserver/src/core/routes/tenants/read.rs
@@ -49,10 +49,12 @@ pub async fn get(
         return list(state, &entry_path, params);
     }
 
-    let entry = state
-        .db
-        .get_entry(&entry_path)?
+    
+    let entry = state.file_service.get_info(&entry_path).await?
         .ok_or_else(|| Error::with_status(StatusCode::NOT_FOUND))?;
+
+    // let stream = state.file_service.get_stream(&entry_path).await?;
+
     let buffer_file = state.db.read_file(&entry.file_id()).await?;
 
     let file_handle = buffer_file.open_file_handle()?;
